Website Builder For NGOs
Non-governmental organizations (NGOs) strive to address some of the most pressing global issues and
promote a better future for all. One of the significant challenges faced by NGOs is to provide education
to the uneducated. Furthermore, the existing online platforms come at a cost, which adds an additional
hurdle. Therefore, develop a website that offers NGOs a tool to create their websites and a platform to
promote their events and activities. Create a solution to assist NGOs in building their online presence
and increasing their outreach to raise awareness and garner volunteers to support for their cause.